backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "dj-rest-auth>=7.0.1",
    "django>=5.2.6",
    "django-allauth[socialaccount]>=65.11.2",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.95.0",
    "filetype>=1.2.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "django-extensions>=4.1",
    "django-browser-reload>=1.20.0",
    "django-debug-toolbar>=6.0.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.13.1",
]
